Use pg instead of ts-postgres and generic-pool dependencies
The pg (node-postgres) seems to be better supported and active than the
other dependencies mentioned.
Also I have troubles to getting it to work lately with Heroku.
Pg has out-of-the-box pool management.
For the DB connection data changed to the single connection string
(DATABASE_URL env variable) as it is auto-updated by Heroku from time to
time as key rotation security policy. In this case we don't need to
worry to update each value if we have it split.